Were made extendable methods for BigInteger and added some tests

+++ Task1_2/Task 1_2/CheckIfNumDevidedBy5.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -8,7 +9,7 @@ namespace Task_1_2
 {
     public static class CheckerIfNumDevidedByNumber
     {
-        public static bool IfCanBeDivByNumber(this int number, int digit)
+		public static bool IfCanBeDivByNumber(this BigInteger number, int digit)
         {
             if (number % digit == 0) return true;
             else return false;
